+++ test/OpenTelemetry.AutoInstrumentation.Tests/Configurations/FileBased/FileBasedOpAmpSettingsTests.cs
@@ -12,7 +12,7 @@ public class FileBasedOpAmpSettingsTests
     [Fact]
     public void LoadFile_GeneralSettings()
     {
-        var serverUrl = "wss://localhost:4318/v1/opamp";
+        var serverUrl = "wss://localhost:4320/v1/opamp";
         var conf = new YamlConfiguration
         {
             OpAmp = new OpAmpConfiguration()