+++ test/OpenTelemetry.AutoInstrumentation.Tests/Configurations/FileBased/Parser/ParserOpAmpTests.cs
@@ -20,7 +20,7 @@ public class ParserOpAmpTests
 
         Assert.Equal("1.0", config.FileFormat);
         Assert.NotNull(config.OpAmp);
-        Assert.Equal("wss://localhost:4318/v1/opamp", config.OpAmp?.ServerUrl);
+        Assert.Equal("wss://localhost:4320/v1/opamp", config.OpAmp?.ServerUrl);
     }
 
     [Fact]
@@ -41,7 +41,7 @@ public class ParserOpAmpTests
     {
         using var envScope = new EnvironmentScope(new Dictionary<string, string?>()
         {
-            { "OTEL_DOTNET_AUTO_OPAMP_SERVER_URL", "wss://localhost:4318/v1/opamp" }
+            { "OTEL_DOTNET_AUTO_OPAMP_SERVER_URL", "wss://localhost:4320/v1/opamp" }
         });
 
         var config = YamlParser.ParseYaml<YamlConfiguration>("Configurations/FileBased/Files/TestOpAmpFileEnvVars.yaml");
@@ -50,6 +50,6 @@ public class ParserOpAmpTests
 
         Assert.Equal("1.0", config.FileFormat);
         Assert.NotNull(config.OpAmp);
-        Assert.Equal("wss://localhost:4318/v1/opamp", config.OpAmp?.ServerUrl);
+        Assert.Equal("wss://localhost:4320/v1/opamp", config.OpAmp?.ServerUrl);
     }
 }