+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Changed
 
+- OpAMP default port changed from 4318 to 4320.
+
 #### Dependency updates
 
 - Following packages updated